@@ -6,7 +6,7 @@ edition = "2021"
 authors = ["10fish <jokefish@live.com>"]
 license-file = "LICENSE"
 repository = "https://github.com/10fish/httprs"
-rust-version = "1.56"
+rust-version = "1.64.0"
 documentation = "https://docs.rs/httprs"
 readme = "README.md"
 keywords = ["http", "server", "static", "tls", "partial_content"]
